refactor(api): extract implant router (B10 start)
