+++ src/squidc5/api/routers/__init__.py
@@ -1,0 +1,1 @@
+"""Domain API routers (split from monolithic routes)."""
