+++ src/squidc5/cli_cmds/__init__.py
@@ -1,0 +1,1 @@
+"""CLI command modules."""
